update Nov. 15

+++ app/models/user.rb
@@ -34,16 +34,29 @@
 #  state                  :string(255)
 #  color_scheme_id        :integer          default(1), not null
 #  notification_level     :integer          default(1), not null
+#  password_expires_at    :datetime
+#  created_by_id          :integer
+#  avatar                 :string(255)
+#  confirmation_token     :string(255)
+#  confirmed_at           :datetime
+#  confirmation_sent_at   :datetime
+#  unconfirmed_email      :string(255)
 #
 
+require 'carrierwave/orm/activerecord'
+require 'file_size_validator'
+
 class User < ActiveRecord::Base
-  devise :database_authenticatable, :token_authenticatable, :lockable,
-         :recoverable, :rememberable, :trackable, :validatable, :omniauthable, :registerable
+  devise :database_authenticatable, :token_authenticatable, :lockable, :async,
+         :recoverable, :rememberable, :trackable, :validatable, :omniauthable, :confirmable, :registerable
 
   attr_accessible :email, :password, :password_confirmation, :remember_me, :bio, :name, :username,
                   :skype, :linkedin, :twitter, :color_scheme_id, :theme_id, :force_random_password,
-                  :extern_uid, :provider, as: [:default, :admin]
-  attr_accessible :projects_limit, :can_create_team, :can_create_group, as: :admin
+                  :extern_uid, :provider, :password_expires_at, :avatar,
+                  as: [:default, :admin]
+
+  attr_accessible :projects_limit, :can_create_group,
+                  as: :admin
 
   attr_accessor :force_random_password
 
@@ -61,23 +74,21 @@ class User < ActiveRecord::Base
   # Namespace for personal projects
   has_one :namespace, dependent: :destroy, foreign_key: :owner_id, class_name: "Namespace", conditions: 'type IS NULL'
 
-  # Namespaces (owned groups and own namespace)
-  has_many :namespaces, foreign_key: :owner_id
-
   # Profile
   has_many :keys, dependent: :destroy
 
   # Groups
-  has_many :groups, class_name: "Group", foreign_key: :owner_id
-
-  # Teams
-  has_many :own_teams,                       dependent: :destroy, class_name: "UserTeam", foreign_key: :owner_id
-  has_many :user_team_user_relationships,    dependent: :destroy
-  has_many :user_teams,                      through: :user_team_user_relationships
-  has_many :user_team_project_relationships, through: :user_teams
-  has_many :team_projects,                   through: :user_team_project_relationships
+  has_many :users_groups, dependent: :destroy
+  has_many :groups, through: :users_groups
+  has_many :owned_groups, through: :users_groups, source: :group, conditions: { users_groups: { group_access: UsersGroup::OWNER } }
 
   # Projects
+  has_many :groups_projects,          through: :groups, source: :projects
+  has_many :personal_projects,        through: :namespace, source: :projects
+  has_many :projects,                 through: :users_projects
+  has_many :created_projects,         foreign_key: :creator_id, class_name: 'Project'
+
+  has_many :snippets,                 dependent: :destroy, foreign_key: :author_id, class_name: "Snippet"
   has_many :users_projects,           dependent: :destroy
   has_many :issues,                   dependent: :destroy, foreign_key: :author_id
   has_many :notes,                    dependent: :destroy, foreign_key: :author_id
@@ -87,10 +98,6 @@ class User < ActiveRecord::Base
   has_many :assigned_issues,          dependent: :destroy, foreign_key: :assignee_id, class_name: "Issue"
   has_many :assigned_merge_requests,  dependent: :destroy, foreign_key: :assignee_id, class_name: "MergeRequest"
 
-  has_many :personal_projects,        through: :namespace, source: :projects
-  has_many :projects,                 through: :users_projects
-  has_many :own_projects,             foreign_key: :creator_id, class_name: 'Project'
-  has_many :owned_projects,           through: :namespaces, source: :projects
 
   #
   # Validations
@@ -101,6 +108,7 @@ class User < ActiveRecord::Base
   validates :extern_uid, allow_blank: true, uniqueness: {scope: :provider}
   validates :projects_limit, presence: true, numericality: {greater_than_or_equal_to: 0}
   validates :username, presence: true, uniqueness: true,
+            exclusion: { in: Gitlab::Blacklist.path },
             format: { with: Gitlab::Regex.username_regex,
                       message: "only letters, digits & '_' '-' '.' allowed. Letter should be first" }
 
@@ -108,8 +116,13 @@ class User < ActiveRecord::Base
 
   validate :namespace_uniq, if: ->(user) { user.username_changed? }
 
+  validates :avatar, file_size: { maximum: 100.kilobytes.to_i }
+
   before_validation :generate_password, on: :create
+  before_validation :sanitize_attrs
+
   before_save :ensure_authentication_token
+
   alias_attribute :private_token, :authentication_token
 
   delegate :path, to: :namespace, allow_nil: true, prefix: true
@@ -118,6 +131,17 @@ class User < ActiveRecord::Base
     after_transition any => :blocked do |user, transition|
       # Remove user from all projects and
       user.users_projects.find_each do |membership|
+        # skip owned resources
+        next if membership.project.owner == user
+
+        return false unless membership.destroy
+      end
+
+      # Remove user from all groups
+      user.users_groups.find_each do |membership|
+        # skip owned resources
+        next if membership.group.last_owner?(user)
+
         return false unless membership.destroy
       end
     end
@@ -131,6 +155,8 @@ class User < ActiveRecord::Base
     end
   end
 
+  mount_uploader :avatar, AttachmentUploader
+
   # Scopes
   scope :admins, -> { where(admin:  true) }
   scope :blocked, -> { with_state(:blocked) }
@@ -140,6 +166,7 @@ class User < ActiveRecord::Base
   scope :not_in_team, ->(team){ where('users.id NOT IN (:ids)', ids: team.member_ids) }
   scope :not_in_project, ->(project) { project.users.present? ? where("id not in (:ids)", ids: project.users.map(&:id) ) : scoped }
   scope :without_projects, -> { where('id NOT IN (SELECT DISTINCT(user_id) FROM users_projects)') }
+  scope :ldap, -> { where(provider:  'ldap') }
 
   scope :potential_team_members, ->(team) { team.members.any? ? active.not_in_team(team) : active  }
 
@@ -147,7 +174,7 @@ class User < ActiveRecord::Base
   # Class methods
   #
   class << self
-    # Devise method overriden to allow sing in with email or username
+    # Devise method overridden to allow sign in with email or username
     def find_for_database_authentication(warden_conditions)
       conditions = warden_conditions.dup
       if login = conditions.delete(:login)
@@ -167,24 +194,28 @@ class User < ActiveRecord::Base
       end
     end
 
-    def create_from_omniauth(auth, ldap = false)
-      gitlab_auth.create_from_omniauth(auth, ldap)
-    end
-
-    def find_or_new_for_omniauth(auth)
-      gitlab_auth.find_or_new_for_omniauth(auth)
-    end
-
-    def find_for_ldap_auth(auth, signed_in_resource = nil)
-      gitlab_auth.find_for_ldap_auth(auth, signed_in_resource)
-    end
-
-    def gitlab_auth
-      Gitlab::Auth.new
-    end
-
     def search query
       where("name LIKE :query OR email LIKE :query OR username LIKE :query", query: "%#{query}%")
+    end
+
+    def by_username_or_id(name_or_id)
+      where('username = ? OR id = ?', name_or_id, name_or_id).first
+    end
+
+    def build_user(attrs = {}, options= {})
+      if options[:as] == :admin
+        User.new(defaults.merge(attrs.symbolize_keys), options)
+      else
+        User.new(attrs, options).with_defaults
+      end
+    end
+
+    def defaults
+      {
+        projects_limit: Gitlab.config.gitlab.default_projects_limit,
+        can_create_group: Gitlab.config.gitlab.default_can_create_group,
+        theme_id: Gitlab.config.gitlab.default_theme
+      }
     end
   end
 
@@ -213,31 +244,27 @@ class User < ActiveRecord::Base
     end
   end
 
-  # Groups where user is an owner
-  def owned_groups
-    groups
-  end
-
-  def owned_teams
-    own_teams
-  end
-
   # Groups user has access to
   def authorized_groups
-    @group_ids ||= (groups.pluck(:id) + authorized_projects.pluck(:namespace_id))
-    Group.where(id: @group_ids)
+    @authorized_groups ||= begin
+                             group_ids = (groups.pluck(:id) + authorized_projects.pluck(:namespace_id))
+                             Group.where(id: group_ids).order('namespaces.name ASC')
+                           end
   end
 
 
   # Projects user has access to
   def authorized_projects
-    @project_ids ||= (owned_projects.pluck(:id) + projects.pluck(:id)).uniq
-    Project.where(id: @project_ids)
+    @authorized_projects ||= begin
+                               project_ids = (personal_projects.pluck(:id) + groups_projects.pluck(:id) + projects.pluck(:id)).uniq
+                               Project.where(id: project_ids).joins(:namespace).order('namespaces.name ASC')
+                             end
   end
 
-  def authorized_teams
-    @team_ids ||= (user_teams.pluck(:id) + own_teams.pluck(:id)).uniq
-    UserTeam.where(id: @team_ids)
+  def owned_projects
+    @owned_projects ||= begin
+                          Project.where(namespace_id: owned_groups.pluck(:id).push(namespace.id)).joins(:namespace)
+                        end
   end
 
   # Team membership in authorized projects
@@ -258,7 +285,7 @@ class User < ActiveRecord::Base
   end
 
   def can_create_project?
-    projects_limit > owned_projects.count
+    projects_limit_left > 0
   end
 
   def can_create_group?
@@ -290,12 +317,12 @@ class User < ActiveRecord::Base
   end
 
   def projects_limit_left
-    projects_limit - owned_projects.count
+    projects_limit - personal_projects.count
   end
 
   def projects_limit_percent
     return 100 if projects_limit.zero?
-    (owned_projects.count.to_f / projects_limit) * 100
+    (personal_projects.count.to_f / projects_limit) * 100
   end
 
   def recent_push project_id = nil
@@ -312,7 +339,7 @@ class User < ActiveRecord::Base
   end
 
   def several_namespaces?
-    namespaces.many?
+    owned_groups.any?
   end
 
   def namespace_id
@@ -325,5 +352,57 @@ class User < ActiveRecord::Base
 
   def tm_of(project)
     project.team_member_by_id(self.id)
+  end
+
+  def already_forked? project
+    !!fork_of(project)
+  end
+
+  def fork_of project
+    links = ForkedProjectLink.where(forked_from_project_id: project, forked_to_project_id: personal_projects)
+
+    if links.any?
+      links.first.forked_to_project
+    else
+      nil
+    end
+  end
+
+  def ldap_user?
+    extern_uid && provider == 'ldap'
+  end
+
+  def accessible_deploy_keys
+    DeployKey.in_projects(self.authorized_projects).uniq
+  end
+
+  def created_by
+    User.find_by_id(created_by_id) if created_by_id
+  end
+
+  def sanitize_attrs
+    %w(name username skype linkedin twitter bio).each do |attr|
+      value = self.send(attr)
+      self.send("#{attr}=", Sanitize.clean(value)) if value.present?
+    end
+  end
+
+  def solo_owned_groups
+    @solo_owned_groups ||= owned_groups.select do |group|
+      group.owners == [self]
+    end
+  end
+
+  def with_defaults
+    User.defaults.each do |k, v|
+      self.send("#{k}=", v)
+    end
+
+    self
+  end
+
+  def can_leave_project?(project)
+    project.namespace != namespace &&
+      project.project_member(self)
   end
 end
